requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Environment & config
python-dotenv==1.0.1

# Video/Audio download
yt-dlp==2024.12.23

# Audio transcription (OpenAI Whisper)
openai-whisper==20240930
# faster-whisper==1.1.0  # Optional: uncomment for GPU-accelerated whisper

# LLM integration (OpenAI-compatible + Anthropic for MiniMax)
openai==1.58.1
anthropic==0.38.0

# HTTP client
httpx==0.28.1

# Utilities
pydantic==2.10.4
